Intigers coverted to floats

+++ Character.cpp
@@ -1,6 +1,6 @@
 #include "Character.h"
 
-Character::Character(const std::string characterName, const unsigned int healthPoints, const unsigned int damagePoints) : name(characterName), DMG(damagePoints)
+Character::Character(const std::string characterName, const float healthPoints, const float damagePoints) : name(characterName), DMG(damagePoints)
 {
 	HP = healthPoints;
 }
@@ -28,7 +28,7 @@ Character Character::parseUnit(std::string fileName)
         std::getline(file, line);
         characterAttributes[2] = line;
 
-        return Character(characterAttributes[0], stoul(characterAttributes[1]), stoul(characterAttributes[2]));
+        return Character(characterAttributes[0], stof(characterAttributes[1]), stof(characterAttributes[2]));
 	}
 	else
 	{
@@ -41,19 +41,19 @@ const std::string Character::getName() const
 	return name;
 }
 
-const unsigned int Character::getHP() const
+const float Character::getHP() const
 {
 	return HP;
 }
 
-const unsigned int Character::getDMG() const
+const float Character::getDMG() const
 {
 	return DMG;
 }
 
-void Character::gotHit(const unsigned int damagePoints)
+void Character::gotHit(const float damagePoints)
 {
-    if (int(HP - damagePoints) > 0)
+    if (float(HP - damagePoints) > 0)
     {
         HP = HP - damagePoints;
     }
